requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Marshmallow==1.2.1
marshmallow-sqlalchemy==1.1.0
marshmallow==3.22.0
Flask-Migrate==4.0.7
mysql-connector-python==8.2.0
python-dotenv==1.0.0
requests==2.31.0
Flask-Limiter==3.5.0
Flask-Caching==2.1.0
python-jose==3.3.0
passlib==1.7.4
flask-swagger==0.2.14
flask-swagger-ui==4.11.1
